Make dbcUsername/Password/ApiUrl optional (only required if plugin is being used)
Because this is an AutoPlugin, it is included in all projects by default. With a complex, multi-project build, it may be annoying to define the task for every project in order to prevent a compile error. This moves the error to become a runtime error when using the plugin, which may be slightly more annoying to users but should be tempered by the README and the error message itself.